[feat] Configure Redux
