+++ package.json
@@ -12,15 +12,19 @@
   "dependencies": {
     "@react-navigation/native": "^6.0.12",
     "@react-navigation/native-stack": "^6.8.0",
+    "@reduxjs/toolkit": "^1.8.5",
     "react": "18.0.0",
     "react-native": "0.69.4",
     "react-native-safe-area-context": "^4.3.3",
-    "react-native-screens": "^3.17.0"
+    "react-native-screens": "^3.17.0",
+    "react-redux": "^8.0.2",
+    "redux": "^4.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
     "@babel/runtime": "^7.12.5",
     "@react-native-community/eslint-config": "^2.0.0",
+    "@redux-devtools/core": "^3.13.1",
     "babel-jest": "^26.6.3",
     "eslint": "^7.32.0",
     "jest": "^26.6.3",
