+++ src/screen/Login/index.js
@@ -3,13 +3,12 @@ import { Image, Text, View } from 'react-native';
 import styles from './styles';
 import TextInput from '../../components/Input/TextInput'
 import Button from '../../components/Button';
+import { connect } from 'react-redux';
 
 const backgroundImageUrl = 'https://cdn.pixabay.com/photo/2017/10/31/19/05/web-design-2906159_1280.jpg'
-const onSubmit = (setIsSigned) => {
-    setIsSigned(true);
-}
 
-const LoginScreen = ({navigation, setIsSigned}) => {
+const LoginScreen = ({login}) => {
+
     return (
         <View style={styles.section}>
             <Image source={{uri: backgroundImageUrl}} style={styles.backgroundImage} ></Image>
@@ -17,11 +16,16 @@ const LoginScreen = ({navigation, setIsSigned}) => {
                 <Text style={styles.title}>ĐĂNG NHẬP</Text>
                 <TextInput placeholder="Please enter your username" style={styles.formInput} />
                 <TextInput placeholder="Please enter your password" style={styles.formInput} secureTextEntry={true} />
-                <Button style={styles.formInput} title={"Submit"} onPress={() => onSubmit(setIsSigned)} />
+                <Button style={styles.formInput} title={"Submit"} onPress={() => login(true)} />
             </View>
         </View>
     );
 }
 
+const mapStateToProps = ({AuthReducer}) => AuthReducer
 
-export default LoginScreen;
+const mapDispatchToProps = (dispatch) => ({
+    login: (isSigned) => dispatch({type: "LOGIN", isSigned})
+})
+
+export default connect(mapStateToProps, mapDispatchToProps)(LoginScreen);
